Replaced command for calling MET binary series_analysis.
Instead of creating a long list of files to pass at the command line, create an ASCII file containing the list of files.
Pass the file name of this ASCII file at the command line instead.
Now this is consistent with how it is done in series_by_lead.py