Fully functional.
Uncommented the section of code responsible for performing the series analysis since testing and development
was being done on the second part of the analysis- generating plots and animations.

The series analysis portion can be replaced by the function create_series_commands() once the bugs are worked out.
Currently create_series_commands() does not create netCDF files that are identical to the originally created
netCDF files (via korn shell script).