Removed unecessary checks and logging.
Don't check for number of variables found in the list of netCDF files, this is irrelevant.
No longer log as a debug message when an IOError exception is thrown when the min.nc and min.txt files
are non-existent.  Removing a file that doesn't exist isn't an error in this context.