Removed extra Params declaration that was causing problems.
When the Params is class is constructed, it parses the command line arguments.  If a second class is constructed the arguments are already parsed.  This meant that when overriding the configuration with "-c <config>" the first instance of params had the overridding params, and the second got the default params.