This fixes tc_pairs_wrapper.py. This fix is temporary to get tc_pairs working the CommandBuilder class.
Do not compare (diff) this commit of tc_pairs_wrapper.py with
the previous commit (which was broken). It will just be confusing.
Rather than trying to figure out what was broken, of which too much
time was spent, this commit is starting from the last known good version.
Which was than modified just enoughto get it working with the
CommandBuilder class.

This version of tc_pairs_wrapper.py was created from the
run_tc_pairs.py commit 270dd6ff in the master branch.

This commit of tc_pairs_wrapper is also relying on the following,
four conf file variables, which had been removed. Again, this is
temporary, to get things working ... the next commit, will use the newer
INIT_BEG and INIT_END variables.

INIT_DATE_BEG = 20141201  ;; DATE must be in YYYYMMDD format
INIT_DATE_END = 20150331  ;;
INIT_HOUR_INC =  6 ;; The increment in integer format
INIT_HOUR_END = 18 ;; Should be a string in HH or HHH format