This commit replaces all references to MET_BUILD_BASE with MET_INSTALL_DIR
MET_BUILD_BASE and RSCRIPTS_BASE should NOT be used anymore.

The only exception is in support of backward compatibility
with the met 6.0 and earlier Rscripts. So for now, while supporting
those releases: MET_BUILD_BASE and RSCRIPTS_BASE only exists
in the metplus_system.conf and tcmpr_plotter_wrapper.py files.