This commit changes the following:
The bulk of the changes in this commit relate to logging
and calling of The external MET and other executables
used in MET+.

logging changes
- All logging settings are configured  in metplus_config/metplus_logging.conf
- LOG_FILENAME is no longer used.

Running executables, the CommandBuilder now has-a CommandRunner
The intention is that all executables will use the CommandRunner.
The command string is passed to the Command Runner which creates
a Runner object, determines where logging is redirected to, and
runs the Runner.

Command line options processing is now unified and handled
in a single function, it was removed from master_metplus.
Using and updated the intended logic that already existed in the
config_metplus.setup() function.

NOTE: BUG FIX to produtil.
It was a 1 line change in ush/produtil/pipeline
The fixed line, call_fcntrl(stream,os.O_NONBLOCK,0,logger)
This issue only impacted Mac AND when adding a file_handler
to the logger for writing to a file.