Bug Fix: feature 101 github #101 Fixed: Logger statements not generating output on JET.
This fix has been tested on Eyewall, Theia, and Jet
The issue had to do with the logger.setLevel(log_level) statement,
in the met_util.py module.
There is a python version dependency, hence the problem was
only occuring on Jet.  Implemented a fix that should work
for all version of python.