Typo Fixes
Corrected the 'getraw_interp' function where references should
have been var_name in the elif statements.

Corrected indent in the 'run_stand_alone' function.
After the function description was added to the function, the existing
indents for the statment blocks, which were eight spaces, had to be
adjusted to match the comment blocks, and was causing an interpreter
error.