v2.0 Bugfix
Typo in variable name in getraw_interp function.
name should be var_name in the if block that is testing for the
existence of the conf variable in the sec argument.

This bug will cause an exception if a conf variable is defined
in the sec section that is used in the opt option that is being
processed by this function. A temp-work-around if the problem is
encountered is to define the conf variable on the config, or dir section.