GitHub Issue #317 remove MET_BASE and replace with MET_INSTALL_DIR
No longer checking for the MET_BASE environment variable.  Now checks for
MET_INSTALL_DIR in the METplus config object and raises an error and exits if
not found.  Now that we are only supporting MET 8.1 and beyond, we no
longer need to check for the RSCRIPTS_BASE or MET_BUILD_BASE environment
variable.  The code corresponding to this check has also been removed to
reduce any confusion and to remove any unecessary code.