Feature 2076 release dropdown menu (#2083)
* test dropdowns

* Update release-notes.rst

Attempting to resolve WARNING

* testing dropdowns

* adding dropdown menus

* testing dropdown

remove from warning section

* dropdown testing

* dropdown menu tests

* dropdown menus

after the warnings section, the dropdown menu only has one space instead of the normal 2 spaces.  If it has 2 spaces, it gets sucked into the warning section.

* spaces with warning

* no spaces for warning

* Update release-notes.rst

warning note has to be indented 1 space.  Originally it was no spaces but then all of the dropdown menus end up in the warning message.  If it's 2 spaces, as is normal, then the warning is indented, which is unnecessary.  1 space is slightly indented and keeps the dropdown menus separate.

---------

Co-authored-by: jprestop <jpresto@ucar.edu>